server/requirements.txt
fastapi
uvicorn
requests
python-multipart
sqlalchemy
PyJWT
